Archive Pandora songs
This changes the usage of the Song model slightly. Instead of creating
a new Song object for every song added to the queue, one is created for
every unique external song identified. The user is recorded as the first
person to add the song. And all users that add the song to the queue are
record as the song's 'lovers'.

Added Resque for doing the background processing. See /resque for the
admin panel (needs to be protected from casual users), Also see README.md
on how to start background workers.